04-rest/0404-rest-boot-client/src/main/java/net/lkrnac/book/eiws/chapter04/boot/RestClientBootApplication.java
package net.lkrnac.book.eiws.chapter04.boot;

import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.context.annotation.Bean;
import org.springframework.web.client.RestTemplate;

@SpringBootApplication
public class RestClientBootApplication {
  public static void main(String... args) {
    SpringApplication.run(RestClientBootApplication.class, args);
  }

  @Bean
  public RestTemplate restTemplate() {
    return new RestTemplate();
  }
}
